backend_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
uuid-utils>=0.10.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
a2a-sdk
google-adk
google-adk-community
litellm
mcp
duckduckgo-mcp-server
uv
pyyaml
structlog
mcp-server-sqlite
mcp-server-git
mcp-server-time
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0